api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
prometheus-client>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
